labellab-flask/requirements.txt
absl-py==0.9.0
alembic==1.4.2
astroid==1.6.6
astunparse==1.6.3
appdirs==1.4.4
backports.functools-lru-cache==1.5
bcrypt==3.1.7
beautifulsoup4==4.8.1
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.2.1
cmake==3.14.3
configparser==3.7.4
coverage==5.2.1
cycler==0.10.0
distlib==0.3.1
enum34==1.1.6
filelock==3.0.12
flake8==3.7.7
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
futures==3.1.1
gast==0.3.3
grpcio==1.29.0
h5py==2.10.0
identify==1.4.29
idna==2.8
importlib-metadata==1.6.1
importlib-resources==3.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
marshmallow==2.21.0
marshmallow-sqlalchemy==0.17.1
mccabe==0.6.1
nose==1.3.7
nodeenv==1.5.0
numpy==1.16.4
oauthlib==3.1.0
onnx==1.6.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.4
pep8==1.7.1
Pillow==7.1.2
protobuf==3.12.2
pre-commit==1.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pylint==1.9.4
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
pyyaml==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
singledispatch==3.4.0.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.8
tb-nightly==2.3.0a20200615
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020061601
tf2onnx==1.6.2
toml==0.10.1
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==16.6.0
Werkzeug==1.0.0
wrapt==1.12.1
zipp==3.1.0